# Optimize MegaAesCtrStream to reduce CPU usage by calling less often into OS socket API