@gpailler 2018-11-11
@GSPP Could you also check which version of MegaApiClient you're using in Duplicati? Because I did so improvements in 1.6.1 one year ago and I noticed Duplicati updated the reference only one month ago https://github.com/duplicati/duplicati/commit/306ccabfef6c2b14ad1026d293d1ebcb9586b2aa#diff-45feddedad3c15dc0db422a7ec69d2aa