@GSPP 2018-11-13
Yes, Duplicati uses that version.

I created a test account and more tests work. Many fail with

> Message: CG.Web.MegaApiClient.ApiException : API response: ResourceExpired

Any idea what that means?