@gpailler 2018-11-15
A test account is not enough to run the tests. You also need a specific folders and files layout in the account (as detailled here https://github.com/gpailler/MegaApiClient/blob/master/MegaApiClient.Tests/Context/AuthenticatedTestContext.cs#L26)

The `API response: ResourceExpired` means a Node Id is not found (and it's related to the layout I mentionned just before).

I think the easiest solution is to update the PR with your changes and the tests will be run by the continuous integration system