@GSPP 2018-11-16
Right, I don't have access to those test files so it would be hard to set up the test file layout correctly. I will rely on the CI tests as you recommended.

When I simply run

        var megaApiClient = new MegaApiClient();
        megaApiClient.Login("...", "...");

I get `ResourceExpired`, too. How can this be?

To give you an update: I have made an initial implementation. It passes the tests that I'm able to run. But when I download a random 10MB file the checksum is invalid. So there's still a bug.

I'm also less optimistic that it is possible to make the whole operation much faster with acceptable amount of work. The bottleneck is the number of AES calls. Each call only works on a single 16 byte block. Apparently, this has a lot of overhead. I implemented a fast path that operates on 64KB blocks. Unfortunately, the way the MAC is defined by Mega the algorithm essentially forces lots of AES calls. I see no way around that. So in order to realize the speed gains that I want I'd need to provide an AES implementation that does not have this much overhead.

I'm quite confident that the AES call overhead is the problem because normally encryption happens at 100s of MB per second. We are getting maybe 5MB/s.

I wanted to try `AesManaged` but it is not available in NetStandard as it seems.