@GSPP 2018-11-16
Profiling a simple 10MB download shows evidence of high call overhead:

![image](https://user-images.githubusercontent.com/12032350/48631199-542b9300-e9be-11e8-91df-8f8f239a4b5a.png)

Ideally, we'd use an AES implementation that simply operates on `byte*`. That would skip all these overheads and multiply total throughput.