@GSPP 2018-11-16
What do you think about copying the source code of `System.Security.Cryptography.RijndaelManagedTransform`? That way we can support fast AES on Netstandard1.3. The class looks fairly self contained when decompiling it.