@gpailler 2018-11-18
@GSPP Did you had time to change the implementation with AesManaged and check the performance improvements?
As a first step, we can use a new implementation for NetStandard 1.6 and .Net Framework and keep the old one for NetStandard 1.3 (I think 1.3 is only required for UWP apps)

About copying/pasting the source code from RijndaelManagedTransform, I need to check the license of the code provided by Microsoft.