@GSPP 2019-01-04
.NET Core is licensed under the MIT license. This means that the code can be used for anything under the condition that a copy of the MIT license is included. Therefore, it seems OK to copy the code and add it to this library.

My tentative plan, **pending your approval**, is to take the algorithm but not the `SymmetricAlgorithm` stuff. The Mega crypto design forces many 16 byte AES invocations. If we want to survive this from a performance standpoint the per-call overhead must be minimal. Therefore, I aim to create a `static void AesEncrypt(ArraySegment<byte> data, byte[] key)` method (or similar). No crypto provider stuff, no initializations, no allocations.

I hope that performance will be good enough if we just switch to this. We should switch for all platforms due to the performance advantages.

I have some fairly complicated code right now that aims to process the Mega data in bigger chunks (as I have explained before). Unfortunately, it is hard to mentally verify and indeed there is still one bug left. I hope that I can just drop that code with the new AES performance gain. If I need to find this bug I might grow gray hair in the process.