@gpailler 2019-01-06
Hello @GSPP 
Yes it seems there is no issue with the License, thanks for the check.
I will be very glad if you can improve the AES encryption part and if you have already some leads. I'm just wondering if copying/pasting some crypto code from .Net Core will not pull a lot of classes and dependencies (that's why I proposed as a first step to improve encryption speed on .Net Standard 1.6 only).

Could you also start your changes from the latest master commit? (I pushed some stability improvements for the tests)