@GSPP 2019-01-06
Yes, I will clean up the commit structure. All of this work should come in the form of one commit on top of master.

I have not looked yet, but I will remove all crypto infrastructure. The core AES stuff should be a bunch of static methods. I have seen AES code before. It is not wildly complicated.

I will get back to you.