@GSPP 2019-01-06
Unfortunately, CoreCLR does not contain a managed code AES implementation. I searched current sources as well as past ones. They delegate to a native library. The `*Managed` classes are kept for compatibility and are now backed by a native implementation. So we cannot take their source code. Microsoft Reference Source does not have a suitable license. I did not find another open source AES implementation for .NET.

My strategy switched to decreasing the overhead on the .NET Framework and keeping compatibility on .NET Core. I personally only care about Duplicati which uses the .NET Framework.

<hr>

Profiling shows that GC overhead is not significant. That's good because it means we do not need to remove temporary allocations of which there are many. The call volume to AES is so high that even small overheads such as virtual calls (e.g. the `CanReuse` check) matter. That's bad news because it means more work for optimizing. Having a custom AES version would solve all of that...

For benchmarking I repeatedly download a 100MB file in a loop. My internet connection is 200Mbit/s but it seems Mega is slower than that. After my changes I am no longer CPU bound so it seems the improvements have payed off.

* Adding `BufferedStream` decreases download duration from 24s to 17s. One CPU core pegged the entire time. I made a custom `BufferedStream` so that it is available on all target frameworks. It also might have less overhead than the BCL stream (who knows). `Stream.Read` is another high volume call target because it is called for each 16 byte chunk.
* Switching to `AesManaged` decreases download duration from 17s to 14s while dropping CPU from 100% of one core to about 60%.
* Optimizing `CachedCryptoTransform` gains a few percent according to my profiler.
* The same for optimizing `IncrementCounter`.

<hr>

Initial profile:

![image](https://user-images.githubusercontent.com/12032350/50739804-599aa200-11e5-11e9-817e-00748825b4b4.png)

The final profile looks like this:

![image](https://user-images.githubusercontent.com/12032350/50739810-65866400-11e5-11e9-858e-7e9e216e0a58.png)

Diff:

![image](https://user-images.githubusercontent.com/12032350/50739819-877fe680-11e5-11e9-9ad0-eea16ab31e08.png)

So it turns out that a custom AES implementation was not needed to reach my performance goals. I am now satisfied with performance. Can you please take a look (https://github.com/gpailler/MegaApiClient/pull/116)?